requirements.txt
tk==0.1.0
python-dotenv==0.20.0
PyYAML==6.0
mysql-connector-python
